Consolidating medDRA tests
